other fixes

+++ src/oofemlib/heap.C
@@ -1,7 +1,7 @@
 #include "heap.h"
 #include <stdlib.h>
 #include <stdio.h>
-#include <math.h>
+#include <mathfem.h>
 
 namespace oofem {
   
@@ -227,8 +227,8 @@ void Heap :: recurse(int row, int pad, int spacing, int S) {
     }
 
     // Calculate the first and last elements on the row
-    int beg = ( pow( 2., ( row - 1 ) ) - 1 );
-    int end = MIN( ( heapCount - 1 ), ( pow(2., row) - 2 ) );
+    int beg = ( pow( 2, ( row - 1 ) ) - 1 );
+    int end = MIN( ( heapCount - 1 ), ( pow(2, row) - 2 ) );
 
     // Newline and padding
     printf("\n");
@@ -251,7 +251,7 @@ void Heap :: printTree() {
     int S = 3;
     int B = 4;
 
-    int nRows = 1 + floor( log2((double)heapCount) );
+    int nRows = 1 + floor( log2(heapCount) );
 
     // Call recurse from the last row
     printf("\n");
